# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rover Control API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RoverStatus {
  connected: boolean;
  host: string;
  port: number;
  latencyMs?: number | null;
  lastCheckedAt: string;
  message?: string | null;
}

export interface RoverCommand {
  /** Hex device ID like "0x400" */
  id: string;
  /** Signed integer value (negative = reverse) */
  value: number;
}

export type RoverMoveAction =
  (typeof RoverMoveAction)[keyof typeof RoverMoveAction];

export const RoverMoveAction = {
  forward: "forward",
  backward: "backward",
  rotate_left: "rotate_left",
  rotate_right: "rotate_right",
  strafe_left: "strafe_left",
  strafe_right: "strafe_right",
  stop: "stop",
  spell: "spell",
} as const;

export interface RoverMove {
  action: RoverMoveAction;
  /** 0-100 speed magnitude */
  speed?: number;
}

export interface RoverCommandResult {
  ok: boolean;
  sent: string[];
  error?: string | null;
}