# lib/api-zod/src/generated/types/roverCommand.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rover Control API
 * OpenAPI spec version: 0.1.0
 */

export interface RoverCommand {
  /** Hex device ID like "0x400" */
  id: string;
  /** Signed integer value (negative = reverse) */
  value: number;
}