# lib/api-zod/src/generated/types/roverCommandResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rover Control API
 * OpenAPI spec version: 0.1.0
 */

export interface RoverCommandResult {
  ok: boolean;
  sent: string[];
  error?: string | null;
}