# lib/api-zod/src/generated/types/roverMove.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rover Control API
 * OpenAPI spec version: 0.1.0
 */
import type { RoverMoveAction } from "./roverMoveAction";

export interface RoverMove {
  action: RoverMoveAction;
  /** 0-100 speed magnitude */
  speed?: number;
}