# lib/api-zod/src/generated/types/roverStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rover Control API
 * OpenAPI spec version: 0.1.0
 */

export interface RoverStatus {
  connected: boolean;
  host: string;
  port: number;
  latencyMs?: number | null;
  lastCheckedAt: string;
  message?: string | null;
}